Move tests and create new sort_index test.
